Extract DistributedValidProducts from OrderCycle
