+++ app/models/order_cycle.rb
@@ -156,7 +156,6 @@ class OrderCycle < ActiveRecord::Base
   end
 
   def distributed_variants
-    # TODO: only used in DistributionChangeValidator, can we remove?
     self.exchanges.outgoing.map(&:variants).flatten.uniq.reject(&:deleted?)
   end
 
@@ -180,10 +179,7 @@ class OrderCycle < ActiveRecord::Base
   # to purchase.
   # This method filters out such products so that the customer cannot purchase them.
   def valid_products_distributed_by(distributor)
-    variants = variants_distributed_by(distributor)
-    products = variants.map(&:product).uniq
-    product_ids = products.reject{ |p| product_has_only_obsolete_master_in_distribution?(p, variants) }.map(&:id)
-    Spree::Product.where(id: product_ids)
+    DistributedValidProducts.new(self, distributor).all
   end
 
   def products
@@ -264,17 +260,6 @@ class OrderCycle < ActiveRecord::Base
   end
 
   private
-
-  # If a product without variants is added to an order cycle, and then some variants are added
-  # to that product, but not the order cycle, then the master variant should not available for customers
-  # to purchase.
-  # This method is used by #valid_products_distributed_by to filter out such products so that
-  # the customer cannot purchase them.
-  def product_has_only_obsolete_master_in_distribution?(product, distributed_variants)
-    product.has_variants? &&
-      distributed_variants.include?(product.master) &&
-      (product.variants & distributed_variants).empty?
-  end
 
   def orders_close_at_after_orders_open_at?
     return if orders_open_at.blank? || orders_close_at.blank?
